Merge authorization into develop

See merge request cse2115/2020-2010/8-corona-proof-room-scheduling/op29-sem58!37

+++ services/rooms/src/main/java/op29sem58/room/controllers/RoomController.java
@@ -2,10 +2,14 @@ package op29sem58.room.controllers;
 
 import java.util.List;
 import java.util.Optional;
+import op29sem58.room.communication.authorization.Authorization;
+import op29sem58.room.communication.authorization.Role;
 import op29sem58.room.entities.Room;
 import op29sem58.room.entities.RoomInfo;
 import op29sem58.room.repositories.RoomRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -14,8 +18,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
-import org.springframework.web.bind.annotation.ResponseBody;
-
 
 /**
  * Controller for Room.
@@ -25,6 +27,8 @@ import org.springframework.web.bind.annotation.ResponseBody;
 public class RoomController {
     @Autowired
     private RoomRepository roomRepository;
+
+    final transient String authHeader = "Authorization";
 
     public RoomRepository getRoomRepository() {
         return roomRepository;
@@ -42,18 +46,14 @@ public class RoomController {
      * @return if the token corresponds to an admin, return "created" message, exception otherwise
      */
     @PostMapping(path = "/createRoom") // Map ONLY POST Requests
-    public @ResponseBody
-    String addNewRoom(@RequestHeader("Authorization") String token, @RequestBody Room r) {
-        token = token.replace("Bearer ", "");
-        boolean isAuthorized = Authorization.authorize(token, "Admin");
-        if (!isAuthorized) {
-            throw new RuntimeException("You do not have the privilege to perform " +
-                    "this action.");
-        } else {
-            roomRepository.saveAndFlush(r);
-            return "Saved room";
-
+    public ResponseEntity<String> addNewRoom(@RequestHeader(authHeader) String token,
+                             @RequestBody Room r) {
+        if (!Authorization.authorize(token, Role.Admin)) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
+
+        roomRepository.saveAndFlush(r);
+        return ResponseEntity.ok("Saved room");
     }
 
     /**
@@ -62,16 +62,12 @@ public class RoomController {
      * @return all Rooms stored in the database
      */
     @GetMapping(path = "/allrooms")
-    public @ResponseBody
-    Iterable<Room> getAllRooms(@RequestHeader("Authorization") String token) {
-        token = token.replace("Bearer ", "");
-        boolean isAuthorized = Authorization.authorize(token, "Admin");
-        if (!isAuthorized) {
-            throw new RuntimeException("You do not have the privilege to perform " +
-                    "this action.");
-        } else {
-            return roomRepository.findAll();
+    public ResponseEntity<Iterable<Room>> getAllRooms(@RequestHeader(authHeader) String token) {
+        if (!Authorization.authorize(token, Role.Admin)) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
+
+        return ResponseEntity.ok(roomRepository.findAll());
     }
 
     /**
@@ -81,15 +77,12 @@ public class RoomController {
      * @return message ("saved" if token corresponds to an admin, exception otherwise)
      */
     @PutMapping(path = "/generateRooms")
-    public @ResponseBody String generateRooms(@RequestHeader("Authorization") String token) {
-        token = token.replace("Bearer ", "");
-        boolean isAuthorized = Authorization.authorize(token, "Admin");
-        if (!isAuthorized) {
-            throw new RuntimeException("You do not have the privilege to perform " +
-                    "this action.");
-        } else {
-            return roomRepository.createInitialRooms();
+    public ResponseEntity<String> generateRooms(@RequestHeader(authHeader) String token) {
+        if (!Authorization.authorize(token, Role.Admin)) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
+
+        return ResponseEntity.ok(roomRepository.createInitialRooms());
     }
 
 
@@ -100,11 +93,10 @@ public class RoomController {
      * @return Room with id 'id'
      */
     @GetMapping(path = "/room/{id}")
-    public @ResponseBody
-    Room getRoom(@PathVariable int id) {
+    public ResponseEntity<Room> getRoom(@PathVariable int id) {
         Optional<Room> room = roomRepository.findById(id);
         if (room.isPresent()) {
-            return room.get();
+            return ResponseEntity.ok(room.get());
         } else {
             throw new RuntimeException("Room not found for the id " + id);
         }
@@ -116,17 +108,24 @@ public class RoomController {
      * @return number of rooms
      */
     @GetMapping(path = "/countRooms")
-    public @ResponseBody
-    Long getNumberOfRooms() {
-        return roomRepository.count();
+    public ResponseEntity<Long> getNumberOfRooms() {
+        return ResponseEntity.ok(roomRepository.count());
     }
 
+    /**
+     * Get rooms with capacity that's at least a set amount.
+     *
+     * @param numOfStudents number of students
+     * @param minPer minimum persons
+     * @param maxPer maximum persons
+     * @return the rooms that meet these conditions
+     */
     @GetMapping(path = "/getRoomsWithCapacityAtLeast/{numOfStudents}/{minPer}/{maxPer}")
-    public @ResponseBody
-    List<RoomInfo> getRoomWithCapAtLeast(@PathVariable Integer numOfStudents,
+    public ResponseEntity<List<RoomInfo>> getRoomWithCapAtLeast(@PathVariable Integer numOfStudents,
                                          @PathVariable Integer minPer,
                                          @PathVariable Integer maxPer) {
-        return roomRepository.getRoomsWithCapacityAtLeast(numOfStudents, minPer, maxPer);
+        return ResponseEntity.ok(roomRepository.getRoomsWithCapacityAtLeast(
+                numOfStudents, minPer, maxPer));
     }
 
     /**
@@ -136,25 +135,36 @@ public class RoomController {
      * @return update message
      */
     @PutMapping(path = "/modifyRoom")
-    public @ResponseBody
-    String editRoom(@RequestBody Room r) {
+    public ResponseEntity<String> editRoom(@RequestHeader(authHeader) String token,
+                                           @RequestBody Room r) {
+        if (!Authorization.authorize(token, Role.Admin)) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
         Optional<Room> room = roomRepository.findById(r.getId());
         if (room.isPresent()) {
             room.get().setName(r.getName());
             room.get().setCapacity(r.getCapacity());
             roomRepository.saveAndFlush(room.get());
-            return "Changed Room";
+            return ResponseEntity.ok("Changed Room");
         } else {
             throw new RuntimeException("Room not found for the id " + r.getId());
         }
     }
 
+    /**
+     * Get the corona capacity of a room.
+     *
+     * @param roomId id of room
+     * @param minPerc min people
+     * @param maxPerc max people
+     * @return the capacity
+     */
     @GetMapping(path = "/getCoronaCapacity/{roomId}/{minPerc}/{maxPerc}")
-    public @ResponseBody
-    Integer coronaCapacity(@PathVariable Integer roomId,
+    public ResponseEntity<Integer> coronaCapacity(@PathVariable Integer roomId,
                              @PathVariable Integer minPerc,
                              @PathVariable Integer maxPerc) {
-        return roomRepository.getCoronaCapacity(roomId, minPerc, maxPerc);
+        return ResponseEntity.ok(roomRepository.getCoronaCapacity(roomId, minPerc, maxPerc));
     }
 
 
@@ -167,12 +177,15 @@ public class RoomController {
      * @return Room with id 'id'
      */
     @DeleteMapping(path = "/deleteRoom/{id}")
-    public @ResponseBody
-    String deleteRoom(@PathVariable int id) {
+    public ResponseEntity<String> deleteRoom(@RequestHeader(authHeader) String token,
+                                             @PathVariable int id) {
+        if (!Authorization.authorize(token, Role.Admin)) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
         Optional<Room> room = roomRepository.findById(id);
         if (room.isPresent()) {
             roomRepository.deleteById(id);
-            return "Deleted Room";
+            return ResponseEntity.ok("Deleted Room");
         } else {
             throw new RuntimeException("Room not found for the id " + id);
         }
